FreeBSD compilation fix.

+++ gdk/gdk_utils.c
@@ -421,8 +421,8 @@ MT_init(void)
 	/* address space (virtual memory) limit */
 	if (getrlimit(RLIMIT_AS, &l) == 0
 	    && l.rlim_cur != RLIM_INFINITY
-	    && l.rlim_cur < GDK_vm_maxsize) {
-		GDK_vm_maxsize = l.rlim_cur;
+	    && (size_t) l.rlim_cur < GDK_vm_maxsize) {
+		GDK_vm_maxsize = (size_t) l.rlim_cur;
 	}
 #endif
 }
